refactor: split session store loader from maintenance
